Fix race condition in waiting for compactor in eunit test.
Monitor waiting for replicator will sometimes fail with noproc
error, because there is a race condition between a running
compactor process and setting up its monitor and waiting on it.

This error appears about once or twice in a 100 runs.
It can be made to appear more often by tweaking 50 and 5 values in:
```
 should_populate_and_compact(RepPid, Source, Target, 50, 5),
```
to something like 1, 20.

This commit fixes race condition by handling noproc.