Allow server admins to retrieve an auth token for any other user
First proposal is a new endpoint /_login_as/<username>, but could be any
other endpoint, even included in /_session, this is just to demonstrate
the feature.

On success, the endpoint returns a JSON object that looks like this:

  {"auth_token": "amFuOjU2M0U0MDREOsGlD2HF6fday16PZGb0vIMDkWCw"}

It returns 404 for nonexistent user or when in admin party mode.

One use-case for this is building delegated authentication mechanisms
in middleware on top of CouchDB (like Oauth), while still being able
to log into CouchDB via it's native auth mechanisms without having
to proxy all requests and their variations, for e.g. replication.

This implementation is for 1.x.x for now (don't judge me), I'll
also supply a 2.x version.