Add recommended docs section
Most people don't program a database and many do not work on
distributed systems.

The idea is to lower the entry barrier to CouchDB and the
concepts behind it. Many possible committers and newcomers are
new to database development and distributed systems.

inspired by https://twitter.com/ag_dubs/status/721874316183396352