Add `delete_after_rename` config parameter
This config parameter allows to specify if the compaction files
must to be deleted right after compaction finished
or we want to pospond this until the database restart.

If latter, the files will be kept in ".delete" directory
renamed to random uuids to avoid collision and cleaned up
during a database start in `couch_file:init_delete_dir/1` call.