Add fast_view filter to changes clustered access
A changes feed can be filtered using view's btree
instead of a database btree, a feature known as fast_view.
This feature not implemented on a clustered interface,
however view filter automatically marked as fast_view
when ddoc has option seq_indexed enabled, which leads to
a crash of filtered replication.

This change allows fast_view filter to be processed
as a regular view filter when used on a clustered
interface.