Really actually fix test for the config listener
Of course now that the handler is tied to the process that calls
listen_for_changes this test was racing the anonymous process that was
starting the handler.

This just turns couch_log_config_listener into a simple gen_server that
sits in couch_log's supervision tree that handles restarting the
listener.

COUCHDB-3069