Update default/local ini files with new settings
A comprehensive grep was run across the source code looking for
references to config:get* functions. This was used to remove any
config settings that no longer are used, and to add new defaults
set in the code explicitly to the ini files. Any newly added lines
were added only as *comments*.

One key change: local.ini still had the [httpd] block for port and
host bindings. As this applies to the "backend" binding and port
now, this section was renamed [chttpd] to ensure new users of 2.0
don't accidentally put their backend port on e.g. port 80.

Fixes COUCHDB-2623.