Fix test suite check for text indexes
This was passing by accident on travis because the $CWD is set
differently so the test would always fail. If however you tried to run
the tests locally then it would fail. Obviously the test is just wrong
since we're moving the mango_cursor_text.erl file in rebar.config, and
not mango_cursor.erl which is required for every index type.

COUCHDB-3143