Fix handling of 413 responses for single document PUT requests
When replicator finds a document which has an attachment size greater than 64k,
or has more than 8 attachments, it switches to a non-batching mode and posts
each document separately using a PUT request with a multipart/related
Content-Type.

Explicitly handle the case when the response to the PUT request is a 413. Skip
the document and dump `doc_write_failures` count, just like in the case of the
413 response for a _bulk_docs POST request.

Jira: COUCHDB-3168