Fix mem3_sync_event_listener unit test
The `should_restart_listener` test was failing because the pattern
supplied to meck was wrong. The argument to
`config_notifier:handle_event/2` is of the pattern `{config_change,
Section, Key, Value, Persist}` but the test was only matching a
four-tuple which I assume was missing the `config_change` atom in the
argument.

COUCHDB-3380