Fix dns timeout for undefined syslog host
On some platforms, inet:getaddr can take several seconds to return when
passed the undefined atom, which is the default value of "syslog_host",
and therefore the value used when running tests. On platforms where it
takes longer to return than the test is allowed to run, the
couch_log_writer_syslog_test will fail with *timed out*.

This short circuits the case when "syslog_host" is undefined, and
removes the race from the test.

COUCHDB-3340