Disabling replication startup jitter in Windows makefile
A similar change has already been made to *nix Makefile already.

This is for the Javascript integration test suite to not timeout
when running in Travis. We don't run Windows tests in Travis but
this should speed things a bit, and it's nice to keep both in
sync.

Jira: COUCHDB-3324