Add jittered sleep during replicator shard scanning
This is bringing back previous code:

https://github.com/apache/couchdb/blob/884cf3e55f77ab1a5f26dc7202ce21771062eae6/src/couch_replicator_manager.erl#L940-L946

This is to avoid a stampede during startup when potentially a large number
shards are found and change feeds have to be opened for all of them at the
same time.

The average jitter value starts at 10 msec for first shard, then goes up to
1 minute for 6000th shard and stays clamped at 1 minute afterwards. (Note:
that's the average, the range is 1 -> 2 * average as this is a uniform
random distribution).

Some sample values:

 * 100 - 1 second
 * 1000 - 10 seconds
 * 6000 and higher - 1 minute

Jira: COUCHDB-3389