Handle non-default _replicator dbs in _scheduler/docs endpoint
Previously _scheduler/docs assumed only the default _replicator db.

To provide consistency and to allow disambiguation between a db named
'db/_replicator' and the document named 'db/_replicator' in the
default replicator db, access to the single document API is changed to
always require the replicator db. That is `/docid` should not be
`/_replicator/docid`.

Now these kinds of paths are accepted after `_scheduler/docs`:

 * `/` : all docs from default _replicator db
 * `/_replicator` : all docs from default replicator db
 * `/other%2f_replicator` : non-default replicator db, urlencoded
 * `/other/_replicator` : non-default replicator db, unencoded
 * `/other%2f_replicator/docid` : doc from a non-default db, urlencoded
 * `/other/_replicator/docid` : doc from a non-default db, db is unencoded

Because `_replicator` is not a valid document ID, it's possible to unambiguously
parse unescaped db paths.

Issue: #506