Choose index based on fields match (#469)
* Choose index based on Prefix and FieldRange

First choose the index with the lowest difference between its 
Prefix and the FieldRanges. If that is equal, then
choose the index with the least number of 
fields in the index. If we still cannot break the tie, 
then choose alphabetically based on ddocId.
Return the first element's Index and IndexRanges.