Skip internal replication if changes already replicated
If minimum checkpointed sequence is greater or equal to source db sequence,
do not start an internal replication task. The typical case is when checkpoint
sequence is equal to the db sequence. Previously replication task was started
always wrote a checkpoint document even if no database changes. This resulted
in a flurry of writes during cluster startup.