Improve JS test harness restartServer() support fn
restartServer() is still erroring out sometimes in Travis/Jenkins. This
PR both bumps the timeout to 15s as well as changes the detection
mechanism for restart to look for the uptime in _system to reset to a
low number.

This PR also removes the eclipsed redundant restartServer() definition
in couch_test_runner.js.

Closes #553