Add new end point to get disk size information for search index
Currently the only way to get the disk size information for the search
index is to use the _search_info end point. But using this point would
lead to opening the search index which is non trivial as lucene has an
overhead for opening the index.

These changes would add a new end point _search_disk_size to get the
disk size information without opening the search index. Sarnie can use
this new end point and avoid opening the search index.

BugzId:87336