Disable unstable JS _show ETag test
The full discussion is in #559, but here is a summary.

Through instrumentation of ddoc_cache and ets_lru I suspect that this is
caused by cache eviction happening after the second GET. While I didn't
instrument exactly where the GET occurs it's clear that it's fairly
late, certainly after the PUT 201 is returned, and likely after the
subsequent GET actually reads from ddoc_cache.

After applying a change to allow me to completely disable the ddoc_cache
(-ddoc_cache max_objects 0 in vm.args) I ran the test on a loop
overnight, and the test never failed (>1000 executions). Previously the
test would fail every 20-30 executions.

TL;DR: we can't guarantee immediate ddoc_cache eviction on a ddoc
update, even for a single .couch file on a single node. (For obvious
reasons we definitely can't guarantee this in a cluster configuration.)

I will document this as a backwards compatibility change in 2.0 and
forward with a separate checkin to couchdb-documentation.

Thanks to @rnewson @janl and @davisp for helping track this one down!

This checkin also includes an improvement to the output when a JS test
fails a notEquals assertion.

Closes #559