Remove another invalid post-2.0 test case (ddoc_cache related)
This test case is failing for the same reason as the failures in #559,
namely a GET on a _show, a PUT to the _show's ddoc to change the _show
function, and a subsequent GET on the same _show that returns a result
that is seemingly outdated. Late ddoc_cache eviction is still the
problem; setting ddoc_cache max_objects to 0 ensures this test always
passes.

Based on the discussion in #559 I am deleting this test as well, direct
on master with approval from @janl and @davisp.