Wait for listener's exit during restart test
To test a config listener's restart we are
deleting event handler in config_event
and then immediately checking event manager.

This creates race when we can be a slightly
early and catch old handler yet to be
removed or slightly late and get
new handler that already been installed.

This patch addresses this by waiting for
the old listener to quit and then waiting
for a new handler to be installed.