Stop couch_index processes on ddoc update
Currently when ddoc is updated, couch_index and couch_index_updater processes
corresponding to the previous version of ddoc will still exist until
all indexing processing initiated by them is done.
When ddoc of a big database is rapidly  modified, this puts a lot
of unnecessary strain on database resources.

With this change, when ddoc is updated:
* all couch_index processes for the previous version of ddoc will be shutdown
* all linked to them couch_index_updater processes will die as well
* all processes waiting for indexing activity to be finished (waiters
    for couch_index:get_status) will receive an immediate reply:
    ddoc_updated. Interactive user requests (view queries) will get response:
    {404, <<"not_found">>, <<"Design document was updated or deleted.">>}

Check if there are ddocs that use the same couch_index process
before closing it on ddoc_updated

1. When opening an index, always add a record {DbName, {DDocId, Sig}} to ?BY_DB.
2. When ddoc_updated, check if there other ddocs in ?BY_DB with the same Sig.
    If there are no, stop couch_index processes.
    If there are other, only remove {DbName, {DDocId, Sig}}
    record from ?BY_DB for this ddoc.