Log OS process I/O
Logging is based on an environment variable:

`COUCHDB_IO_LOG_DIR`

If set, logs will go to that directory.

Logs are per `couch_os_process` Erlang process. There are 3 files saved for
each process:

```
<unixtimestamp>_<erlangpid>.in.log : Input, data coming from the proess
<unixtimestamp>_<erlangpid>.out.log : Output, data going to the process
<unixtimestamp>_<erlangpid>.meta : Error reason
```

Log files are saved as named (visible) files only if an error occurs. If there
is no error, disk space will still be used as long the process is alive. But as
soon as it exists, file will be unlinked and space will be reclaimed.

Issue: #551